Finegrained access control for snapshot restore
This commit implements finefrained access control for the snapshot restore operation.
A user, who wants to restore an index from a snapshot does not only need the restore right (cluster:admin/snapshot/restore) but also
the right to create (indices:admin/create) and write (indices:data/write/index) to to the destination index and on every type within the destination index.

This feature is configurable in elasticsearch.yml config file.

The feature builds on and was roughly outlined in https://github.com/floragunncom/search-guard/pull/245

WIP: not yet feature complete, one test case is not yet implemented (and thous only added as comment)